+++ modules/services/model/src/main/java/au/com/shiftyjelly/pocketcasts/models/db/helper/ListenedCategory.kt
@@ -5,10 +5,19 @@ import au.com.shiftyjelly.pocketcasts.models.entity.Podcast
 data class ListenedCategory(
     val episodeId: String,
     val numberOfPodcasts: Int,
+    val numberOfEpisodes: Int,
     val totalPlayedTime: Long,
     val category: String,
     val mostListenedPodcastId: String,
     val mostListenedPodcastTintColor: Int,
 ) {
     fun toPodcast() = Podcast(uuid = mostListenedPodcastId, tintColorForLightBg = mostListenedPodcastTintColor)
+
+    fun simplifiedCategoryName() = when (category) {
+        "Health & Fitness" -> "Health"
+        "Kids & Family" -> "Family"
+        "Religion & Spirituality" -> "Spirituality"
+        "Society & Culture" -> "Culture"
+        else -> category
+    }
 }