+++ modules/services/repositories/src/main/java/au/com/shiftyjelly/pocketcasts/repositories/endofyear/EndOfYearManagerImpl.kt
@@ -96,20 +96,20 @@ class EndOfYearManagerImpl @Inject constructor(
         val stories = mutableListOf<Story>()
 
         stories.add(StoryIntro())
-        listeningTime?.let { stories.add(StoryListeningTime(it, topPodcasts.takeLast(3))) }
-        if (listenedCategories.isNotEmpty()) {
-            stories.add(StoryListenedCategories(listenedCategories))
-            stories.add(StoryTopListenedCategories(listenedCategories))
-        }
         if (listenedNumbers.numberOfEpisodes > 1 && listenedNumbers.numberOfPodcasts > 1) {
             stories.add(StoryListenedNumbers(listenedNumbers, topPodcasts))
         }
         if (topPodcasts.isNotEmpty()) {
-            stories.add(StoryTopPodcast(topPodcasts.first()))
+            stories.add(StoryTopPodcast(topPodcasts = topPodcasts))
             if (topPodcasts.size > 1) {
                 stories.add(StoryTopFivePodcasts(topPodcasts.take(5)))
             }
         }
+        if (listenedCategories.isNotEmpty()) {
+            stories.add(StoryTopListenedCategories(listenedCategories))
+            stories.add(StoryListenedCategories(listenedCategories))
+        }
+        listeningTime?.let { stories.add(StoryListeningTime(it, topPodcasts.takeLast(3))) }
         longestEpisode?.let { stories.add(StoryLongestEpisode(it)) }
         stories.add(StoryEpilogue())